AMQ Health Check
